fix incorrect previous routine sets and reps
The problem here was that when you do alternating routines each day (for
example bodyweight routine one day, stretching the next) then the
previous reps and sets information was not correctly shown.

The code prior to this patch searched for the previous routine in the
realm repository that had a start time less then now, with a unique id
different to the current routine. This found routines that were of any
type, the bodyweight routine, mobility or stretching. If the current
routine was not the same type then the "previous sets" information was
wrong since it was indexed into a different routine type and referenced
a completely different exercise.

This patch adds a check to say find prior routines with the same "title"
as the currently chosen one, which correctly pulls out the previous
bodyweight/mobility/stretching routine matching the current routine
type. After that, the set and rep information is correctly read.