Maybe fix the "wrong time" when restoring problem
The problem: a timed exercise is next, and there is a rest timer running
because you have logged the current exercise. You click to the next
exercise while the timer is running. When the timer stops, the current
seconds for the timed exercise do not show your chosen value but instead
show the default 1:00 time.

The fix is to add a restoreView method that resets the timer value.