Continue using the last used weight value
Get the weight of the last set from the previous routine and use that as
the next weight for the first set of this routine.

This is a bit of a hack to avoid adding a change weight button to the
main UI. Instead, you'll need to modify a set to add weight, but then
the other sets will at least not reset back to 0kg. Each subsequent set
uses the weight from last time in this session's set.

e.g. last time: 6+5kg, 6+5kg, 6+5kg
then this time, first set would automatically use 6+5kg
if you change it to 6+7kg then set 2 would be 6+7kg too.

This means you don't have to faff around changing weight so often, it
should do the right thing. And it saves cluttering up the UI with
more buttons to adjust the default weight.